Refactoring-regular checkin. - removed global context - created config holding class with more type information - expanded diagnostic context - moved package structure and renamed some classes
