+++ src/main/java/com/elastic/support/PostProcessor.java
@@ -1,4 +1,4 @@
-package com.elastic.support.scrub;
+package com.elastic.support;
 
 public interface PostProcessor {
    public String process(String input);
