+++ src/main/java/com/elastic/support/diagnostics/commands/BaseQueryCmd.java
@@ -1,45 +1,52 @@
 package com.elastic.support.diagnostics.commands;
 
-import com.elastic.support.diagnostics.DiagnosticInputs;
+import com.elastic.support.config.DiagnosticInputs;
 import com.elastic.support.diagnostics.chain.Command;
-import com.elastic.support.diagnostics.chain.DiagnosticContext;
 import com.elastic.support.diagnostics.chain.GlobalContext;
 import com.elastic.support.rest.RestExec;
 import com.elastic.support.util.SystemProperties;
-import org.apache.http.HttpHost;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 public abstract class BaseQueryCmd implements Command {
 
     protected final Logger logger = LogManager.getLogger(BaseQueryCmd.class);
 
-    public void runQuery(Map.Entry<String, String> entry, DiagnosticContext context) {
+    public void runQuery(String host, int port, String scheme, String queryName, String query, String tempDir) {
 
         RestExec restExec = GlobalContext.getRestExec();
         DiagnosticInputs diagnosticInputs = GlobalContext.getDiagnosticInputs();
 
-        String queryName = entry.getKey();
-        String query = entry.getValue();
 
         logger.debug(": now processing " + queryName + ", " + query);
 
         logger.info("Currently running query: {}", query);
 
         String fileName = buildFileName(queryName, context);
-        HttpHost httpHost = new HttpHost(
-                diagnosticInputs.getHost(),
-                diagnosticInputs.getPort(),
-                diagnosticInputs.getScheme()
+
+
       );
 
-        restExec.execConfiguredDiagnosticQuery(query, fileName, httpHost);
+        restExec.execConfiguredDiagnosticQuery(query, fileName, host, port, scheme);
     }
 
-    protected String buildFileName(String queryName, DiagnosticContext context) {
+    protected void runQueries(Set<Map.Entry<String, String>> entries, String host, int port, String scheme, String queryName){
+
+        String tempDir = context.getTempDir();
+
+        for (Map.Entry<String, String> entry : entries) {
+            String queryName = entry.getKey();
+            String query = entry.getValue();
+            runQuery(queryName, query, tempDir);
+        }
+
+    }
+
+    protected String buildFileName(String queryName) {
 
         List textFileExtensions = (List) GlobalContext
                 .getConfig().get("textFileExtensions");
@@ -50,7 +57,7 @@ public abstract class BaseQueryCmd implements Command {
         } else {
             ext = ".json";
         }
-        String fileName = context.getTempDir() + SystemProperties.fileSeparator + queryName + ext;
+        String fileName = GlobalContext.getDiagnosticInputs().getTempDir() + SystemProperties.fileSeparator + queryName + ext;
 
         return fileName;
