+++ src/main/java/com/elastic/support/diagnostics/commands/HostIdentifierCmd.java
@@ -1,7 +1,6 @@
 package com.elastic.support.diagnostics.commands;
 
-import com.elastic.support.diagnostics.Constants;
-import com.elastic.support.diagnostics.Diagnostic;
+import com.elastic.support.config.Constants;
 import com.elastic.support.diagnostics.chain.Command;
 import com.elastic.support.diagnostics.chain.DiagnosticContext;
 import com.elastic.support.diagnostics.chain.GlobalContext;
