+++ src/main/java/com/elastic/support/diagnostics/commands/NodeManifestCmd.java
@@ -1,6 +1,6 @@
 package com.elastic.support.diagnostics.commands;
 
-import com.elastic.support.diagnostics.DiagnosticInputs;
+import com.elastic.support.config.DiagnosticInputs;
 import com.elastic.support.diagnostics.chain.Command;
 import com.elastic.support.diagnostics.chain.DiagnosticContext;
 import com.elastic.support.diagnostics.chain.GlobalContext;
