+++ src/main/java/com/elastic/support/diagnostics/commands/RunLogstashQueriesCmd.java
@@ -1,6 +1,5 @@
 package com.elastic.support.diagnostics.commands;
 
-import com.elastic.support.diagnostics.Diagnostic;
 import com.elastic.support.diagnostics.chain.DiagnosticContext;
 import com.elastic.support.diagnostics.chain.GlobalContext;
 import com.elastic.support.util.JsonYamlUtils;
