+++ src/test/resources/diags-test.yml
@@ -50,104 +50,31 @@ retries:
   shards: 3
 
 rest-calls:
-  # count: 36
+
   common:
     alias: "/_alias?pretty&human"
-    cat_aliases: "/_cat/aliases?v"
-    cat_allocation: "/_cat/allocation?v"
-    cat_count: "/_cat/count"
-    cat_fielddata: "/_cat/fielddata?v"
-    cat_health: "/_cat/health?v"
     cat_indices: "/_cat/indices?v"
-    cat_master: "/_cat/master?format=json"
-    cat_nodes: "/_cat/nodes?v&h=n,m,i,r,d,hp,rp,cpu,load_1m,load_5m,load_15m,nodeId"
-    cat_pending_tasks: "/_cat/pending_tasks?v"
-    cat_segments: "/_cat/segments?v"
-    cat_recovery: "/_cat/recovery?v"
-    cat_shards: "/_cat/shards?v"
-    cat_thread_pool: "/_cat/thread_pool?v"
-    cluster_health: "/_cluster/health?pretty"
-    cluster_pending_tasks: "/_cluster/pending_tasks?pretty&human"
-    cluster_settings: "/_cluster/settings?pretty&flat_settings"
-    cluster_state: "/_cluster/state?pretty&human"
-    cluster_stats: "/_cluster/stats?pretty&human"
-    fielddata: "/_cat/fielddata?format=json&bytes&pretty"
-    fielddata_stats: "/_nodes/stats/indices/fielddata?pretty=true&fields=*"
-    indices_stats: "/_stats?pretty&human"
-    mapping: "/_mapping?pretty"
-    master: "/_cat/master?format=json"
-    nodes: "/_nodes?pretty&human"
-    nodes_hot_threads: "/_nodes/hot_threads?threads=10000"
-    nodes_stats: "/_nodes/stats?pretty&human"
-    plugins: "/_cat/plugins?format=json"
-    recovery: "/_recovery?pretty&human&detailed=true"
-    segments: "/_segments?pretty&human&verbose"
-    settings: "/_settings?pretty&human"
-    shards: "/_cat/shards?format=json&bytes=b&pretty"
-    templates: "/_template?pretty"
-    version: "/"
-    watcher_stats: "/_watcher/stats/_all"
-    watcher_stack: "/_watcher/stats?emit_stacktraces=true"
 
   versions:
-    # count 1
     major-1:
       minor-0:
         licenses: "/_licenses"
-    # total 37
 
     major-2:
-      # count: 8 - 1 override = 7
       minor-0:
         cat_nodeattrs: "/_cat/nodeattrs?v&h=node,id,pid,host,ip,port,attr,value"
-        cat_repositories: "/_cat/repositories?v"
-        count: "/_count"
-        licenses: "/_license?pretty"  #override
-        security_users: "/_shield/user?pretty"
-        security_roles: "/_shield/role?pretty"
-        shard_stores: "/_shard_stores?pretty"
-        tasks: "/_tasks?pretty&human&detailed=true"
-    #total 44
 
     major-5:
-      # count 11 - 3 override = 9
       minor-0:
         allocation_explain: "/_cluster/allocation/explain?pretty"
-        allocation_explain_disk: "/_cluster/allocation/explain?include_disk_info=true&pretty"
-        fielddata_stats: "/_nodes/stats/indices/fielddata?level=shards&pretty=true&fields=*" #override
-        ml_anomaly_detectors: "/_xpack/ml/anomaly_detectors?pretty"
-        ml_datafeeds: "/_xpack/ml/datafeeds?pretty"
-        ml_stats: "/_xpack/ml/anomaly_detectors/_stats?pretty"
-        pipelines: "/_ingest/pipeline/*?pretty&human"
-        security_users: "/_xpack/security/user?pretty" #override
-        security_roles: "/_xpack/security/role?pretty" #override
-        security_role_mappings: "/_xpack/security/role_mapping?pretty"
-        xpack: "/_xpack/usage?pretty&human"
-    # total: 52
+      minor-2:
+        cat_indices: "_cat/indices?v&s=index"
 
     major-6:
       minor-0:
-        # count 2
         nodes_usage: "/_nodes/usage"
-        remote_cluster_info: "/_remote/info"
-      #total 54
-
-      minor-3:
-        # count 2
-        rollup_jobs: "/_xpack/rollup/job/_all"
-        rollup_caps: "/_xpack/rollup/data/_all"
-      #total 56
-
-      minor-4:
-        # count 1
-        cluster_settings_defaults: "/_cluster/settings?include_defaults&pretty&flat_settings"
-      #total 57
-
       minor-5:
-        # count 2
-        rollup_index_caps: "/*/_xpack/rollup/data"
         security_priv: "/_xpack/security/privilege?pretty"
-      # total 59
 
 elastic-threads:
   nodes: "/_nodes?pretty&human"
